requirements.txt
pandas
seaborn
lftk
pyenchant